requirements-backend-only.txt
# Backend Requirements - Lightweight Web Service
# This file contains ONLY dependencies needed for the web backend
# All ML dependencies have been moved to ml-service/

# Django Core
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.7.0
django-filter==24.3
django-extensions==3.2.3
django-timezone-field==7.1

# Authentication & Security
PyJWT==2.10.1

# Database
psycopg2-binary==2.9.10

# Task Queue
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
amqp==5.3.1
billiard==4.2.1
kombu==5.4.2
vine==5.1.0

# Utilities
python-dateutil==2.9.0.post0
python-crontab==3.2.0
pytz==2025.1
cron-descriptor==1.4.5
emoji==2.14.1

# HTTP Client (for calling ML service)
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
chardet==5.2.0
idna==3.10

# Payments
razorpay==1.4.2

# PDF Generation & Analytics Visualization
reportlab==4.3.1
pillow==11.1.0
xhtml2pdf==0.2.16
matplotlib==3.9.0

# Server
gunicorn==21.2.0

# Template Engine
Jinja2==3.1.5
MarkupSafe==3.0.2

# CLI
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6

# Other
sqlparse==0.5.3
tzdata==2025.1

# ==========================================
# REMOVED ML DEPENDENCIES (Now in ml-service):
# ==========================================
# joblib==1.4.2
# pandas==2.2.3
# scikit-learn==1.6.1
# scipy==1.15.2
# nltk==3.9.1
# ==========================================
# NOTE: numpy is now present as a matplotlib dependency (~12MB)
# This is acceptable as it's needed for analytics visualization
# The key removal is the 600MB+ ML training libraries (sklearn, scipy, pandas, nltk)
# ==========================================
# regex==2024.11.6
# Cython==3.0.12
# filelock==3.17.0
# fsspec==2025.2.0
# mpmath==1.3.0
# networkx==3.4.2
# prompt_toolkit==3.0.50
# protobuf==5.29.3
# six==1.17.0
# SpeechRecognition==3.14.1
# stanza==1.10.1
# sympy==1.13.1
# threadpoolctl==3.5.0
# tqdm==4.67.1
# typing_extensions==4.12.2
# wcwidth==0.2.13
# wheel==0.45.1
# setuptools==75.8.0
